+++ src/main/java/me/shadowalzazel/mcodyssey/enchantments/OdysseyEnchantments.kt
@@ -4,6 +4,8 @@ import me.shadowalzazel.mcodyssey.enchantments.armor.*
 import me.shadowalzazel.mcodyssey.enchantments.melee.*
 import me.shadowalzazel.mcodyssey.enchantments.other.MirrorForce
 import me.shadowalzazel.mcodyssey.enchantments.ranged.AlchemyArtillery
+import me.shadowalzazel.mcodyssey.enchantments.ranged.LuckyDraw
+import me.shadowalzazel.mcodyssey.enchantments.ranged.Rend
 import me.shadowalzazel.mcodyssey.enchantments.utility.GildedPower
 import org.bukkit.enchantments.Enchantment
 import java.util.*
@@ -14,7 +16,7 @@ object OdysseyEnchantments {
     // Utility
     val GILDED_POWER: Enchantment = GildedPower
     // Enchantments
-    val ALCHEMYARTILLERY: Enchantment = AlchemyArtillery
+    val ALCHEMY_ARTILLERY: Enchantment = AlchemyArtillery
     val BACKSTABBER: Enchantment = Backstabber
     val BANE_OF_THE_ILLAGER: Enchantment = BaneOfTheIllager
     val BANE_OF_THE_SEA: Enchantment = BaneOfTheSea
@@ -30,8 +32,10 @@ object OdysseyEnchantments {
     val GRAVITY_WELL: Enchantment = GravityWell
     val GUARDING_STRIKE: Enchantment = GuardingStrike
     val HEMORRHAGE: Enchantment = Hemorrhage
+    val LUCKY_DRAW: Enchantment = LuckyDraw
     val MIRROR_FORCE: Enchantment = MirrorForce
     val POTION_BARRIER: Enchantment = PotionBarrier
+    val REND: Enchantment = Rend
     val SPOREFUL: Enchantment = Sporeful
     val SQUIDIFY: Enchantment = Squidify
     val VENGEFUL: Enchantment = Vengeful
@@ -39,8 +43,8 @@ object OdysseyEnchantments {
     val WHIRLWIND: Enchantment = Whirlwind
 
     // Set of all enchantments
-    val enchantmentSet = setOf(GILDED_POWER, ALCHEMYARTILLERY, BANE_OF_THE_ILLAGER, BANE_OF_THE_SEA, BANE_OF_THE_SWINE, BUZZY_BEES, COWARDICE, DECAYING_TOUCH, DOUSE,
-        EXPLODING, FREEZING_ASPECT, FROG_FRIGHT, FRUITFUL_FARE, GRAVITY_WELL, GUARDING_STRIKE, HEMORRHAGE, MIRROR_FORCE, POTION_BARRIER, SPOREFUL, SQUIDIFY, VOID_STRIKE, WHIRLWIND)
+    val enchantmentSet = setOf(GILDED_POWER, ALCHEMY_ARTILLERY, BANE_OF_THE_ILLAGER, BANE_OF_THE_SEA, BANE_OF_THE_SWINE, BUZZY_BEES, COWARDICE, DECAYING_TOUCH, DOUSE,
+        EXPLODING, FREEZING_ASPECT, FROG_FRIGHT, FRUITFUL_FARE, GRAVITY_WELL, GUARDING_STRIKE, HEMORRHAGE, LUCKY_DRAW, MIRROR_FORCE, POTION_BARRIER, REND, SPOREFUL, SQUIDIFY, VOID_STRIKE, WHIRLWIND)
 
     // Register
     fun register() {
