+++ src/main/java/me/shadowalzazel/mcodyssey/enchantments/ranged/LuckyDraw.kt
@@ -1,16 +1,15 @@
 package me.shadowalzazel.mcodyssey.enchantments.ranged
 
-import me.shadowalzazel.mcodyssey.enchantments.OdysseyEnchantments
 import me.shadowalzazel.mcodyssey.enchantments.utility.OdysseyEnchantmentWrapper
 import org.bukkit.Material
 import org.bukkit.enchantments.Enchantment
 import org.bukkit.inventory.ItemStack
 
-object AlchemyArtillery : OdysseyEnchantmentWrapper("alchemyartillery", "Alchemy Artillery", 3) {
+object LuckyDraw : OdysseyEnchantmentWrapper("luckydraw", "Lucky Draw", 3) {
 
     override fun conflictsWith(other: Enchantment): Boolean {
         return when (other) {
-            OdysseyEnchantments.REND -> {
+            ARROW_INFINITE, MENDING -> {
                 true
             }
             else -> {
@@ -21,7 +20,7 @@ object AlchemyArtillery : OdysseyEnchantmentWrapper("alchemyartillery", "Alchemy
 
     override fun canEnchantItem(item: ItemStack): Boolean {
         return when (item.type) {
-            Material.ENCHANTED_BOOK, Material.CROSSBOW -> {
+            Material.ENCHANTED_BOOK, Material.BOW -> {
                 true
             }
             else -> {
